XMLHttpRequest/send-accept.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - Accept</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/accept.php", false)
        client.send(null)
        assert_equals(client.responseText, "*/*")
        client.open("GET", "resources/accept.php", false)
        client.setRequestHeader("Accept", "x-something/vague, text/html5")
        client.send(null)
        assert_equals(client.responseText, "x-something/vague, text/html5")
      })
    </script>
  </body>
</html>
